Limit CR3 previews to JPEG only
(cherry picked from commit 40e5021577025562ca7eb083486af3bc612d04d9)

+++ include/exiv2/bmffimage.hpp
@@ -121,10 +121,10 @@ namespace Exiv2
         void parseCr3Preview(DataBuf &data,
                              std::ostream &out,
                              bool bTrace,
-                             uint16_t width_offset,
-                             uint16_t height_offset,
+                             uint32_t width_offset,
+                             uint32_t height_offset,
                              uint32_t size_offset,
-                             uint16_t relative_position);
+                             uint32_t relative_position);
         //@}
 
         //! @name Manipulators
